Widgets refactored & Title center problem has been solved
