+++ lib/presentation/ui/screens/setting_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_talk/presentation/ui/utility/assets_path.dart';
+import 'package:flutter_talk/presentation/ui/widgets/setting_option_tile.dart';
 
 class SettingScreen extends StatefulWidget {
   const SettingScreen({super.key});
@@ -135,49 +136,6 @@ class _SettingScreenState extends State<SettingScreen> {
           ],
         ),
       ),
-    );
-  }
-}
-
-class SettingOptionTile extends StatelessWidget {
-  const SettingOptionTile({
-    super.key,
-    required this.icon,
-    required this.title,
-    this.subtitle,
-  });
-
-  final IconData icon;
-  final String title;
-  final String? subtitle;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      leading: CircleAvatar(
-          radius: 22,
-          backgroundColor: const Color(0xffF2F8F7),
-          child: Icon(
-            icon,
-            size: 24,
-            color: const Color(0xff797C7B),
-          )),
-      title: Text(
-        title,
-        style: const TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w600,
-            color: Color(0xff000E08)),
-      ),
-      subtitle: subtitle != null
-          ? Text(
-              subtitle.toString(),
-              style: const TextStyle(
-                fontWeight: FontWeight.w500,
-                fontSize: 12,
-              ),
-            )
-          : null,
     );
   }
 }
